feature(asymmetric): extracted tests to separate files
